+++ src/main/kotlin/com/example/demo/HistoryApplication.kt
@@ -1,0 +1,11 @@
+package com.example.demo
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class HistoryApplication
+
+fun main(args: Array<String>) {
+    runApplication<HistoryApplication>(*args)
+}
