+++ src/main/kotlin/com/example/demo/history/application/exception/HistoryNotFoundException.kt
@@ -1,0 +1,5 @@
+package com.example.demo.history.application.exception
+
+import javassist.NotFoundException
+
+class HistoryNotFoundException(override val message: String) : NotFoundException(message)
