+++ src/main/kotlin/com/example/demo/history/application/input/UpdateHistoryInput.kt
@@ -1,0 +1,13 @@
+package com.example.demo.history.application.input
+
+import com.fasterxml.jackson.annotation.JsonFormat
+import java.time.LocalDate
+
+
+data class UpdateHistoryInput(
+    val title: String,
+    val content: String,
+    val improvements: String?,
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd") val startDate: LocalDate,
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd") val endDate: LocalDate?
+)
