+++ src/main/kotlin/com/example/demo/history/application/service/FindHistoryUseCase.kt
@@ -1,0 +1,9 @@
+package com.example.demo.history.application.service
+
+import com.example.demo.history.domain.entity.History
+
+interface FindHistoryUseCase {
+    fun findById(id: String): History
+
+    fun findAll(): Iterable<History>
+}
