+++ src/main/kotlin/com/example/demo/history/application/service/RemoveHistoryUseCase.kt
@@ -1,0 +1,7 @@
+package com.example.demo.history.application.service
+
+import com.example.demo.history.domain.entity.History
+
+interface RemoveHistoryUseCase {
+    fun remove(history: History)
+}
