+++ src/main/kotlin/com/example/demo/history/application/service/UpdateHistoryUseCase.kt
@@ -1,0 +1,7 @@
+package com.example.demo.history.application.service
+
+import com.example.demo.history.domain.entity.History
+
+interface UpdateHistoryUseCase {
+    fun update(id: String, history: History): History
+}
