+++ src/main/kotlin/com/example/demo/history/domain/repository/HistoryRepository.kt
@@ -1,0 +1,9 @@
+package com.example.demo.history.domain.repository
+
+import com.example.demo.history.domain.entity.History
+import org.springframework.data.repository.CrudRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface HistoryRepository : CrudRepository<History, String> {
+}
