--- src/main/kotlin/com/example/history/HistoryApplication.kt
@@ -1,11 +1,0 @@
-package com.example.history
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class HistoryApplication
-
-fun main(args: Array<String>) {
-    runApplication<HistoryApplication>(*args)
-}
