--- src/main/kotlin/com/example/history/application/exception/HistoryNotFoundException.kt
@@ -1,5 +1,0 @@
-package com.example.history.application.exception
-
-import javassist.NotFoundException
-
-class HistoryNotFoundException(override val message: String) : NotFoundException(message)
