--- src/main/kotlin/com/example/history/application/input/UpdateHistoryInput.kt
@@ -1,13 +1,0 @@
-package com.example.history.application.input
-
-import com.fasterxml.jackson.annotation.JsonFormat
-import java.time.LocalDate
-
-
-data class UpdateHistoryInput(
-    val title: String,
-    val content: String,
-    val improvements: String?,
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd") val startDate: LocalDate,
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd") val endDate: LocalDate?
-)
