--- src/main/kotlin/com/example/history/application/service/RemoveHistoryUseCase.kt
@@ -1,7 +1,0 @@
-package com.example.history.application.service
-
-import com.example.history.domain.entity.History
-
-interface RemoveHistoryUseCase {
-    fun remove(history: History)
-}
