--- src/main/kotlin/com/example/history/application/service/SaveHistoryUseCase.kt
@@ -1,7 +1,0 @@
-package com.example.history.application.service
-
-import com.example.history.domain.entity.History
-
-interface SaveHistoryUseCase {
-    fun save(history: History): History
-}
