--- src/main/kotlin/com/example/history/application/service/UpdateHistoryUseCase.kt
@@ -1,7 +1,0 @@
-package com.example.history.application.service
-
-import com.example.history.domain.entity.History
-
-interface UpdateHistoryUseCase {
-    fun update(id: String, history: History): History
-}
