--- src/main/kotlin/com/example/history/domain/repository/HistoryRepository.kt
@@ -1,9 +1,0 @@
-package com.example.history.domain.repository
-
-import com.example.history.domain.entity.History
-import org.springframework.data.repository.CrudRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface HistoryRepository : CrudRepository<History, String> {
-}
