--- src/main/kotlin/com/example/util/DateTimeUtil.kt
@@ -1,10 +1,0 @@
-package com.example.util
-
-import java.text.SimpleDateFormat
-import java.time.LocalDate
-
-fun getDateByFormat(date: LocalDate, format: String): String {
-    val simpleDateFormat = SimpleDateFormat(format)
-
-    return simpleDateFormat.format(date)
-}
