+++ src/test/kotlin/com/example/demo/history/HistoryApplicationTests.kt
@@ -1,0 +1,16 @@
+package com.example.demo.history
+
+import org.mockito.Mockito
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class HistoryApplicationTests {
+}
+
+/**
+ * @link {https://withhamit.tistory.com/138}
+ */
+fun <T> any(): T {
+    Mockito.any<T>()
+    return null as T
+}
